Fix TypedMoneyAddressRegistry and allow for custom registries
