+++ src/test/kotlin/xyz/block/moneyaddress/typed/MobileTypedMoneyAddressTest.kt
@@ -1,7 +1,7 @@
 package xyz.block.moneyaddress.typed
 
 import xyz.block.moneyaddress.KES
-import xyz.block.moneyaddress.TypedMoneyAddressTest.Companion.makeMoneyAddress
+import xyz.block.moneyaddress.MoneyAddressExamples.Companion.makeMoneyAddress
 import xyz.block.moneyaddress.UNRECOGNIZED_CURRENCY
 import xyz.block.moneyaddress.typed.MobileMoneyAddress.Companion.toMobileMoneyAddress
 import kotlin.test.Test
